requirements.txt
deepagents==0.3.7
langchain==1.2.6
langchain-openai==1.0.3
langgraph==1.0.3
streamlit==1.47.1
python-docx==1.1.0
pywin32==311
tavily-python==0.7.19
requests==2.32.5
